[style] css 파일 분리 및 페이지 내 별도 컴포넌트 분리
