+++ src/components/IssueDetail/Markdown.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+import { IssueDetailProps } from '../../types/IssueDetailTyes';
+
+interface MarkdownProps {
+  issueDetailInfo: IssueDetailProps;
+}
+
+export default function Markdown({ issueDetailInfo }: MarkdownProps) {
+  return (
+    <ReactMarkdown remarkPlugins={[remarkGfm]}>
+      {issueDetailInfo.body}
+    </ReactMarkdown>
+  );
+}
