+++ src/pages/IssueDetail/IssueDetail.tsx
@@ -5,63 +5,17 @@ import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import { githubApi } from '../../api/GithubIssueApi';
 import { useIssueContext } from '../../context/IssueContext';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import { css } from '@emotion/react';
-
-const flexAlign = css`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const imgFlexAlign = css`
-  ${flexAlign}
-`;
-
-const imgSize = css`
-  width: 50px;
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-  border-radius: 50%;
-`;
-
-const issueTitleArea = css`
-  display: flex;
-  align-items: center;
-  font-size: 20px;
-`;
-
-const issueTextArea = css`
-  ${issueTitleArea}
-  font-size: 16px;
-`;
-
-const issueTitle = css`
-  font-weight: 700;
-  margin-left: 10px;
-  white-space: nowrap;
-`;
-
-const issueText = css`
-  ${issueTitle}
-  font-weight: 400;
-  margin: 10px 0 10px 10px;
-`;
-
-const commentStyle = css`
-  white-space: nowrap;
-`;
-
-const line = css`
-  width: 100%;
-  height: 1px;
-  background-color: #c0c0c0;
-  margin: 10px 0;
-`;
+import {
+  commentStyle,
+  flexAlign,
+  imgFlexAlign,
+  imgSize,
+  issueText,
+  issueTextArea,
+  issueTitleArea,
+  line,
+} from './IssueDetailCSS';
+import Markdown from '../../components/IssueDetail/Markdown';
 
 export default function IssueDetail() {
   const [isLoading, setIsLoading] = useState(false);
@@ -99,7 +53,7 @@ export default function IssueDetail() {
     <div>
       {isLoading && <div>로딩중</div>}
       {!isLoading && (
-        <>
+        <div>
           <div css={flexAlign}>
             <div>
               <div css={imgFlexAlign}>
@@ -127,11 +81,8 @@ export default function IssueDetail() {
             <div css={commentStyle}>코멘트: {issueDetailInfo.comments}</div>
           </div>
           <div css={line} />
-          <ReactMarkdown
-            children={issueDetailInfo.body}
-            remarkPlugins={[remarkGfm]}
-          />
-        </>
+          <Markdown issueDetailInfo={issueDetailInfo} />
+        </div>
       )}
     </div>
   );
